+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Support for SDL1 rendering using surface blitting
 * Mouse position recalibration to match game cursor position
 * Add savestate slot to input editor
+* Old autosaves are removed
+* Autosave configuration window
 
 ### Changed